Artikel Verändern

+++ src/domain/exceptions/UngueltigerPreisException.java
@@ -2,7 +2,7 @@ package domain.exceptions;
 
 public class UngueltigerPreisException extends RuntimeException {
 
-    public UngueltigerPreisException(float preis) {
+    public UngueltigerPreisException(double preis) {
         super("Ungültiger Preis: " + preis + ". Der Preis darf nicht negativ sein.");
     }
 }
